Factor out a ton of common code in the tests, reduce imports, set default log level to the now-standard info
