+++ Tests/AsyncKitTests/EventLoop+ConcurrencyTests.swift
@@ -1,9 +1,8 @@
 import XCTest
 import AsyncKit
 import NIOCore
-import NIOPosix
 
-final class EventLoopConcurrencyTests: XCTestCase {
+final class EventLoopConcurrencyTests: AsyncKitTestCase {
     func testGroupMakeFutureWithTask() throws {
         @Sendable
         func getOne() async throws -> Int {
@@ -14,25 +13,6 @@ final class EventLoopConcurrencyTests: XCTestCase {
         }
         
         XCTAssertEqual(try expectedOne.wait(), 1)
-    }
-
-    var group: EventLoopGroup!
-    var eventLoop: EventLoop { self.group.any() }
-
-    override func setUpWithError() throws {
-        try super.setUpWithError()
-        self.group = MultiThreadedEventLoopGroup(numberOfThreads: 1)
-    }
-
-    override func tearDownWithError() throws {
-        try self.group.syncShutdownGracefully()
-        self.group = nil
-        try super.tearDownWithError()
-    }
-
-    override class func setUp() {
-        super.setUp()
-        XCTAssertTrue(isLoggingConfigured)
     }
 }
