+++ Tests/AsyncKitTests/EventLoopFutureQueueTests.swift
@@ -2,9 +2,8 @@ import NIOConcurrencyHelpers
 import AsyncKit
 import XCTest
 import NIOCore
-import NIOPosix
 
-final class EventLoopFutureQueueTests: XCTestCase {
+final class EventLoopFutureQueueTests: AsyncKitTestCase {
     func testQueue() throws {
         let queue = EventLoopFutureQueue(eventLoop: self.eventLoop)
         var numbers: [Int] = []
@@ -189,25 +188,6 @@ final class EventLoopFutureQueueTests: XCTestCase {
         )
 
         XCTAssertEqual(error.description, "previousError(nope)")
-    }
-
-    var group: EventLoopGroup!
-    var eventLoop: EventLoop { self.group.any() }
-
-    override func setUpWithError() throws {
-        try super.setUpWithError()
-        self.group = MultiThreadedEventLoopGroup(numberOfThreads: 1)
-    }
-
-    override func tearDownWithError() throws {
-        try self.group.syncShutdownGracefully()
-        self.group = nil
-        try super.tearDownWithError()
-    }
-
-    override class func setUp() {
-        super.setUp()
-        XCTAssertTrue(isLoggingConfigured)
     }
 }
