+++ Tests/AsyncKitTests/Future+OptionalTests.swift
@@ -1,9 +1,8 @@
 import AsyncKit
 import XCTest
 import NIOCore
-import NIOPosix
 
-final class FutureOptionalTests: XCTestCase {
+final class FutureOptionalTests: AsyncKitTestCase {
     func testOptionalMap() throws {
         let future = self.eventLoop.makeSucceededFuture(Optional<Int>.some(1))
         let null = self.eventLoop.makeSucceededFuture(Optional<Int>.none)
@@ -53,24 +52,5 @@ final class FutureOptionalTests: XCTestCase {
     
     func multiply(_ a: Int, _ b: Int?) -> EventLoopFuture<Int?> {
         return self.group.any().makeSucceededFuture(b == nil ? nil : a * b!)
-    }
-
-    var group: EventLoopGroup!
-    var eventLoop: EventLoop { self.group.any() }
-
-    override func setUpWithError() throws {
-        try super.setUpWithError()
-        self.group = MultiThreadedEventLoopGroup(numberOfThreads: 1)
-    }
-
-    override func tearDownWithError() throws {
-        try self.group.syncShutdownGracefully()
-        self.group = nil
-        try super.tearDownWithError()
-    }
-
-    override class func setUp() {
-        super.setUp()
-        XCTAssertTrue(isLoggingConfigured)
     }
 }
